Conflict system prepared for final playtesting.
Created player death state and modified board text based on playtest feedback,Changed climbing material to hlp players identify it as an interactable, Made the door check for sicknessbar activity to determine whether it stays locked or unlocked yada yada. Made the scene functional to the best of my abilities.